requirements.txt
mysql-connector-python
pymongo
PyMySQL